2024.07.19 1.小程序联调成功，配置yml 2.完成OAuth验证 3.user端setmeal/category完成 4.练习mybatisplus 5.练习setmealmapper.xml中使用left join

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -66,7 +66,7 @@ public class DishController {
      *
      */
     @GetMapping("/list")
-    public Result<List<Dish>> list( Integer categoryId){
+    public Result<List<Dish>> list(Long categoryId){
         List<Dish> list = dishService.list(categoryId);
         return Result.success(list);
     }
