+++ sky-server/src/main/java/com/sky/mapper/DishMapper.java
@@ -27,5 +27,5 @@ public interface DishMapper extends BaseMapper<Dish> {
 
 
     @Select("select * from dish where category_id = #{categoryId}")
-    List<Dish> list(Integer categoryId);
+    List<Dish> list(Long categoryId);
 }
